test(it): add integration test target
