+++ .env.repo
@@ -1,17 +1,23 @@
 # Space seperated list of features that needs to be tested on their own
 # Needs to be quoted if multiple entries are used
-EXCLUSIVE_FEATURES_TEST="stub_client"
+EXCLUSIVE_FEATURES_TEST="stub_client stub_server"
+
 # Comma seperated list of features that needs to be enabled for test/ build/ release
-PACKAGE_TEST_FEATURES=test_util
+PACKAGE_UT_FEATURES=test_util
+PACKAGE_IT_FEATURES=all_resources
 PACKAGE_BUILD_FEATURES=all_resources
 PACKAGE_RELEASE_FEATURES=default
+
+# Create the 'dev' image with stub_backends feature enabled if you want to run the server
+# without the need of spinning up all dependend services.
+DOCKER_DEV_FEATURES=
 
 # PostgreSQL driver connection settings
 PG__USER=svc_storage
 PG__DBNAME=arrow
 PG__HOST=cockroachdb
 PG__PORT=26257
-PG__SSLMODE=require
+PG__SSL_MODE=Require # Must start with capital to match the SslMode Enum
 
 # Database SSL paths
 DB_CA_CERT=/cockroach/ssl/certs/ca.crt
