+++ client-grpc/Cargo.toml
@@ -20,7 +20,7 @@ include = [
 ]
 
 [features]
-dev       = ["mock", "all_resources"]
+dev       = ["mock", "all_resources", "svc-storage"]
 test_util = ["mock", "stub_backends", "all_resources"]
 # Supported resources
 adsb = ["any_resource"]
@@ -54,8 +54,10 @@ vertipad = ["any_resource", "vertiport"]
 vertiport = ["any_resource"]
 # Will add a 'mock' module for the enabled resources, providing access to mock data generation functions
 mock = ["chrono", "rand", "geo"]
-# Will use a stubbed server connection, only use for tests!
+# Will use a stubbed backend connection, only use for tests!
 stub_backends = ["svc-storage/stub_server", "lib-common/grpc_mock", "tower"]
+# Pass through stub_server feature for our server code
+stub_server = ["stub_backends"]
 # Will implement stub functions for the client, only use for tests!
 stub_client = [
   "lazy_static",
