+++ client-grpc/tests/resources/group_vehicle.rs
@@ -119,4 +119,13 @@ pub async fn scenario(
         .ids
         .iter()
         .any(|id| id == &vehicle_ids[vehicle_ids.len() - 4]));
+
+    // Remove all linked for group
+    let result = link_client
+        .unlink(Id {
+            id: groups.list[0].id.clone(),
+        })
+        .await;
+    println!("{:?}", result);
+    assert!(result.is_ok());
 }
