+++ client-grpc/tests/resources/group_vertipad.rs
@@ -108,4 +108,13 @@ pub async fn scenario(
         .ids
         .iter()
         .any(|id| id == &vertipad_ids[vertipad_ids.len() - 4]));
+
+    // Remove all linked for group
+    let result = link_client
+        .unlink(Id {
+            id: groups.list[0].id.clone(),
+        })
+        .await;
+    println!("{:?}", result);
+    assert!(result.is_ok());
 }
