+++ docker-compose.yml
@@ -2,15 +2,24 @@
 ## DO NOT EDIT!
 ## This file was provisioned by Terraform
 ## File origin: https://github.com/Arrow-air/tf-github/tree/main/src/templates/rust-svc/docker-compose.yml
-version: '3.6'
+
+version: '3.8'
+
+configs:
+  log4rs:
+    file: log4rs.yaml
+  dot-env:
+    file: .env
+
 volumes:
   cockroachdb:
   cockroachdb-ssl:
 
 services:
   cockroachdb-init:
-    container_name: ${DOCKER_NAME}-example-cockroachdb-init
+    container_name: ${DOCKER_NAME}-cockroachdb-init
     image: cockroachdb/cockroach:latest-v22.1
+    env_file: .env
     volumes:
       - type: volume
         source: cockroachdb
@@ -26,8 +35,9 @@ services:
     command: ["/scripts/cockroachdb-init.sh"]
 
   cockroachdb:
-    container_name: ${DOCKER_NAME}-example-cockroachdb
+    container_name: ${DOCKER_NAME}-cockroachdb
     image: cockroachdb/cockroach:latest-v22.1
+    env_file: .env
     depends_on:
       cockroachdb-init:
         condition: service_completed_successfully
@@ -53,58 +63,39 @@ services:
     command: start-single-node --certs-dir=/cockroach/ssl/certs --advertise-addr=cockroachdb
 
   web-server:
-    container_name: ${DOCKER_NAME}-example-server
-    image: ${PACKAGE_NAME}:latest
-    ports:
-      - ${HOST_PORT_REST}:${DOCKER_PORT_REST}
-      - ${HOST_PORT_GRPC}:${DOCKER_PORT_GRPC}
-    healthcheck:
-      test: ["CMD", "grpc_health_probe", "-addr", "localhost:${DOCKER_PORT_GRPC}"]
-      interval: 2s
-      timeout: 1s
-      retries: 3
-      start_period: 30s
+    extends:
+      file: docker-compose-base.yml
+      service: web-server
     volumes:
-      - type: bind
-        source: log4rs.yaml
-        target: ${LOG_CONFIG}
-      - type: bind
-        source: logs/
-        target: /logs/
-      - type: bind
-        source: .env
-        target: /.env
       - type: volume
         source: cockroachdb-ssl
         target: /cockroach/ssl
         read_only: true
     command: --rebuild-psql=true  --init-psql=true
-    environment:
-      - RUST_BACKTRACE
     depends_on:
       cockroachdb:
         condition: service_healthy
 
   example:
+    extends:
+      file: docker-compose-base.yml
+      service: example
     links:
       - cockroachdb
-      - web-server
-    depends_on:
-      web-server:
-        condition: service_healthy
-    container_name: ${DOCKER_NAME}-example
-    image: ${RUST_IMAGE_NAME}:${RUST_IMAGE_TAG}
+
+  ut-coverage:
+    extends:
+      file: docker-compose-base.yml
+      service: ut-coverage
+
+  it-coverage:
+    extends:
+      file: docker-compose-base.yml
+      service: it-coverage
+    links:
+      - cockroachdb
     volumes:
-      - type: bind
-        source: "${SOURCE_PATH}/"
-        target: "/usr/src/app"
-      - type: bind
-        source: "${SOURCE_PATH}/.cargo/registry"
-        target: "/usr/local/cargo/registry"
-    environment:
-      - SERVER_HOSTNAME
-      - SERVER_PORT_GRPC
-      - SERVER_PORT_REST
-      - EXAMPLE_TARGET
-      - RUST_BACKTRACE
-    command: cargo run --manifest-path "${CARGO_MANIFEST_PATH}" --example "${EXAMPLE_TARGET}"
+      - type: volume
+        source: cockroachdb-ssl
+        target: /cockroach/ssl
+        read_only: true
