+++ server/src/main.rs
@@ -1,8 +1,6 @@
 //! Main function starting the gRPC server and initializing dependencies.
 
 use log::info;
-use svc_storage::postgres::init::{create_db, recreate_db};
-use svc_storage::postgres::init_psql_pool;
 use svc_storage::*;
 
 /// Main entry point: starts gRPC Server on specified address and port
@@ -20,22 +18,21 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     info!("(main) Server startup.");
 
-    info!("(main) Running database initialization.");
-    init_psql_pool().await?;
-
     // Allow options for psql init or and/ or recreation
     // locally: cargo run -- --init-psql true
     let args = Cli::parse();
     if let Some(rebuild_psql) = args.rebuild_psql {
         if rebuild_psql {
             info!("(main) Found argument [rebuild_psql]. Rebuilding now...");
-            recreate_db().await?;
+            #[cfg(not(feature = "stub_backends"))]
+            svc_storage::postgres::init::recreate_db().await?;
             info!("(main) PSQL Rebuild completed.");
         }
     } else if let Some(init_psql) = args.init_psql {
         if init_psql {
             info!("(main) Found argument [init_psql]. Creating database schema now...");
-            create_db().await?;
+            #[cfg(not(feature = "stub_backends"))]
+            svc_storage::postgres::init::create_db().await?;
             info!("(main) PSQL Database creation completed.");
         }
     }
