+++ server/src/postgres/simple_resource_linked.rs
@@ -1,6 +1,6 @@
 //! Psql Simple resource Traits
 
-use super::get_psql_pool;
+use super::get_psql_client;
 use super::{util::*, ArrErr, PsqlData, PsqlField, PsqlFieldSend};
 use crate::grpc::server::ValidationResult;
 use crate::grpc::{GrpcDataObjectType, GrpcField};
@@ -90,7 +90,7 @@ where
                 }
             }
             None => {
-                let client = get_psql_pool().get().await?;
+                let client = get_psql_client().await?;
                 let stmt = client.prepare_cached(&query).await?;
                 match client.execute(&stmt, &ref_params[..]).await {
                     Ok(rows) => {
@@ -139,7 +139,7 @@ where
         psql_debug!("(create) [{}].", insert_sql);
         psql_debug!("(create) [{:?}].", &params);
 
-        let client = get_psql_pool().get().await?;
+        let client = get_psql_client().await?;
         client
             .execute(insert_sql, &params[..])
             .await
@@ -215,7 +215,7 @@ where
         psql_debug!("(update) [{}].", update_sql);
         psql_debug!("(update) [{:?}].", &params);
 
-        let client = get_psql_pool().get().await?;
+        let client = get_psql_client().await?;
         client.execute(update_sql, &params[..]).await?;
 
         //TODO(R4): flush shared memcache for this resource when memcache is implemented
@@ -284,7 +284,7 @@ where
             definition.psql_table,
             id
         );
-        let client = get_psql_pool().get().await?;
+        let client = get_psql_client().await?;
 
         let query = format!(
             r#"UPDATE "{}" SET "deleted_at" = NOW() WHERE "{}" = $1"#,
@@ -329,7 +329,7 @@ where
             definition.psql_table,
             id
         );
-        let client = get_psql_pool().get().await?;
+        let client = get_psql_client().await?;
         let query = format!(
             r#"DELETE FROM "{}" WHERE "{}" = $1"#,
             definition.psql_table, id_col
